feat(data-table): enhance pagination and cell rendering in DataTable
- Updated DataTablePagination component to improve layout and styling, including adjustments to padding and text size for better readability.
- Refined AutoCell component to enhance rendering logic for various cell types, including badges and currency, with improved default handling and formatting.
- Streamlined the handling of status variants in AutoCell for better clarity and maintainability.
- Removed deprecated properties and ensured consistent use of props across components.

These changes enhance the user experience and visual consistency of the data table components.
