+++ kit/dapp/src/components/data-table/data-table.tsx
@@ -224,7 +224,10 @@ export function DataTable<TData, CParams extends Record<string, unknown>>({
   // Bulk actions state and handlers (after table creation)
   const isBulkActionsEnabled = bulkActions?.enabled ?? false;
   const selectedRowIds = Object.keys(currentState.rowSelection ?? {}).filter(
-    (key) => currentState.rowSelection?.[key as keyof typeof currentState.rowSelection] === true
+    (key) =>
+      currentState.rowSelection?.[
+        key as keyof typeof currentState.rowSelection
+      ] === true
   );
   const selectedRows = useMemo(() => {
     if (!isBulkActionsEnabled || selectedRowIds.length === 0) return [];
